Fixed caching bug in imaging service
git-svn-id: https://wush.net/svn/mindalliance@1882 baad322d-9929-0410-88f0-f92e8ff3e1bd

+++ channels/trunk/src/main/java/com/mindalliance/channels/imaging/DefaultImagingService.java
@@ -22,6 +22,7 @@ import java.awt.image.Kernel;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.net.URLEncoder;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -113,8 +114,8 @@ public class DefaultImagingService implements ImagingService {
             BufferedImage image;
             if ( isUploadedFileDocument( url ) ) {
                 image = ImageIO.read( getUploadedImageFile( url ) );
-            } else if ( isFileDocument( url )) {
-                image = ImageIO.read( new File( url) );
+            } else if ( isFileDocument( url ) ) {
+                image = ImageIO.read( new File( url ) );
             } else {
                 image = ImageIO.read( new URL( url ) );
             }
@@ -203,19 +204,19 @@ public class DefaultImagingService implements ImagingService {
      */
     public void deiconize( ModelObject modelObject ) {
         File iconFile = getIconFile( modelObject );
-        iconFile.delete();
-        File squaredIconFile = getIconFile( modelObject );
-        if ( squaredIconFile.exists() ) squaredIconFile.delete();
+        if ( iconFile.exists() ) iconFile.delete();
+        iconFile = getIconFile( modelObject );
+        if ( iconFile.exists() ) iconFile.delete();
         for ( int i = 1; i < ICON_HEIGHTS.length; i++ ) {
             iconFile = getIconFile( modelObject, i );
-            iconFile.delete();
+            if ( iconFile.exists() )iconFile.delete();
         }
     }
 
     /**
      * {@inheritDoc}
      */
-    public String getIconPath( ModelObject modelObject ) {
+    public String getModelObjectIconsPath( ModelObject modelObject ) {
         File iconFile = getIconFile( modelObject );
         if ( iconFile.exists() ) {
             String path = iconFile.getAbsolutePath();
@@ -227,7 +228,7 @@ public class DefaultImagingService implements ImagingService {
             if ( modelObject.hasImage() ) {
                 boolean success = iconize( modelObject.getImageUrl(), modelObject );
                 if ( success )
-                    return getIconPath( modelObject );
+                    return getModelObjectIconsPath( modelObject );
                 else
                     return null;
             } else {
@@ -242,13 +243,21 @@ public class DefaultImagingService implements ImagingService {
     public String getSquareIconUrl( ModelObject modelObject ) {
         File squareIconFile = getSquareIconFile( modelObject );
         if ( squareIconFile.exists() ) {
-            return "icons"
-                    + File.separator
-                    + getIconsSubDirName( modelObject)
-                    + "|"
-                    + squareIconFile.getName();
+            try {
+                String encodedPath = URLEncoder.encode(
+                        getIconsPath( modelObject )
+                                + File.separator
+                                + squareIconFile.getName()
+                        , "UTF-8" );
+                return "icons"
+                        + File.separator
+                        + encodedPath;
+            } catch ( Exception e ) {
+                LOG.warn( "Failed to get icon url", e );
+                return null;
+            }
         } else {
-            String iconPath = getIconPath( modelObject );
+            String iconPath = getModelObjectIconsPath( modelObject );
             if ( iconPath != null ) {
                 boolean success = squarify( iconPath + ".png", modelObject );
                 if ( success )
@@ -258,10 +267,6 @@ public class DefaultImagingService implements ImagingService {
             }
             return null;
         }
-    }
-
-    private String getIconsSubDirName( ModelObject modelObject ) {
-        return modelObject.getClass().getSimpleName().toLowerCase();
     }
 
     private void createNumberedIcons( BufferedImage resized, int width, ModelObject modelObject ) throws IOException {
@@ -340,32 +345,32 @@ public class DefaultImagingService implements ImagingService {
     }
 
     private File getIconFile( ModelObject modelObject ) {
-        String path = getIconDirectory( modelObject ).getAbsolutePath()
+        String path = getIconsDirectory( modelObject ).getAbsolutePath()
                 + File.separator
-                + sanitize( modelObject.getName() )
+                + sanitizeFileName( modelObject.getName() )
                 + ".png";
         return new File( path );
     }
 
     private File getIconFile( ModelObject modelObject, int index ) {
-        String path = getIconDirectory( modelObject ).getAbsolutePath()
+        String path = getIconsDirectory( modelObject ).getAbsolutePath()
                 + File.separator
-                + sanitize( modelObject.getName() )
+                + sanitizeFileName( modelObject.getName() )
                 + index
                 + ".png";
         return new File( path );
     }
 
     private File getSquareIconFile( ModelObject modelObject ) {
-        String path = getIconDirectory( modelObject ).getAbsolutePath()
+        String path = getIconsDirectory( modelObject ).getAbsolutePath()
                 + File.separator
-                + sanitize( modelObject.getName() )
+                + sanitizeFileName( modelObject.getName() )
                 + "_squared.png";
         return new File( path );
     }
 
 
-    private String sanitize( String fileName ) {
+    private String sanitizeFileName( String fileName ) {
         if ( File.separator.equals( "\\" ) ) {
             return fileName.replaceAll( "\\\\", "" );
         } else {
@@ -374,27 +379,57 @@ public class DefaultImagingService implements ImagingService {
     }
 
     /**
+     * Return a "directory-safe" equivalent name.
+     *
+     * @param name original name
+     * @return safe version
+     */
+    public static String uriToDirName( String name ) {
+        return name.replaceAll( "\\W", "_" );
+    }
+
+
+    /**
      * Get the location of the generated icons for a given model object.
      *
      * @param modelObject a model object
      * @return a directory
      */
-    private File getIconDirectory( ModelObject modelObject ) {
+    private File getIconsDirectory( ModelObject modelObject ) {
         try {
-            File iconDir = iconDirectory.getFile();
-            if ( !iconDir.exists() ) {
-                iconDir.mkdir();
-            }
-            String subDirName = getIconsSubDirName( modelObject );
-            File subDir = new File( iconDir.getAbsolutePath() + File.separator + subDirName );
-            if ( !subDir.exists() ) {
-                subDir.mkdir();
-            }
-            return subDir;
+            String moIconsSubDirName = getIconsPath( modelObject );
+            return new File( moIconsSubDirName );
         } catch ( IOException e ) {
             throw new RuntimeException( e );
         }
     }
+
+    private File getBaseIconDirectory() throws IOException {
+        File iconDir = iconDirectory.getFile();
+        if ( !iconDir.exists() ) {
+            iconDir.mkdir();
+        }
+        return iconDir;
+    }
+
+    private String getIconsPath( ModelObject modelObject ) throws IOException {
+        File iconDir = getBaseIconDirectory();
+        String planSpecificDirName = iconDir.getAbsolutePath()
+                + File.separator
+                + uriToDirName( User.current().getPlan().getVersionUri() );
+        File subDir = new File( planSpecificDirName );
+        if ( !subDir.exists() ) {
+            subDir.mkdir();
+        }
+        String moDirName = modelObject.getClass().getSimpleName().toLowerCase();
+        String moDirPath = subDir.getAbsolutePath() + File.separator + moDirName;
+        File moSubDir = new File( moDirPath );
+        if ( !moSubDir.exists() ) {
+            moSubDir.mkdir();
+        }
+        return moDirPath;
+    }
+
 
     /**
      * Find icon name for given model object.
@@ -404,7 +439,7 @@ public class DefaultImagingService implements ImagingService {
      * @return a string
      */
     public String findIconName( ModelObject modelObject, String imagesDirName ) {
-        String iconName = getIconPath( modelObject );
+        String iconName = getModelObjectIconsPath( modelObject );
         if ( iconName == null ) {
             if ( modelObject instanceof Actor ) {
                 iconName = imagesDirName + '/'
@@ -429,11 +464,11 @@ public class DefaultImagingService implements ImagingService {
             if ( part.getActor().isType() ) {
                 Actor knownActor = part.getKnownActualActor( queryService );
                 if ( knownActor != null ) {
-                    iconName = getIconPath( knownActor );
+                    iconName = getModelObjectIconsPath( knownActor );
                 }
             }
             if ( iconName == null ) {
-                iconName = getIconPath( part.getActor() );
+                iconName = getModelObjectIconsPath( part.getActor() );
             }
             if ( iconName == null ) {
                 iconName = this + "/" + ( part.isSystem() ? "system" : "person" );
@@ -442,18 +477,18 @@ public class DefaultImagingService implements ImagingService {
             Actor knownActor = part.getKnownActualActor( queryService );
             boolean onePlayer = knownActor != null;
             if ( onePlayer ) {
-                iconName = getIconPath( knownActor );
+                iconName = getModelObjectIconsPath( knownActor );
                 if ( iconName == null ) {
                     iconName = imagesDirName + "/" + ( part.isSystem() ? "system" : "person" );
                 }
             } else {
-                iconName = getIconPath( part.getRole() );
+                iconName = getModelObjectIconsPath( part.getRole() );
                 if ( iconName == null ) {
                     iconName = imagesDirName + "/" + ( part.isSystem() ? "system" : "role" );
                 }
             }
         } else if ( part.getOrganization() != null ) {
-            iconName = getIconPath( part.getOrganization() );
+            iconName = getModelObjectIconsPath( part.getOrganization() );
             if ( iconName == null ) {
                 iconName = imagesDirName + "/" + "organization";
             }
